feat(forum): implement inline file upload for non-image blocks in forum editor
